+++ negative_training_sampler/cli.py
@@ -6,17 +6,26 @@ from negative_training_sampler.negative_training_balancer import balance_trainin
 
 @click.command()
 @click.option("-i",
-              "label_file",
+              "--label-file",
+              'label_file',
               required=True,
               type=click.Path(),
               help="Input bed file with labeled regions")
 @click.option("-r",
-              "genome_file",
+              "--reference-file",
+              'reference_file',
               required=True,
               type=click.Path(),
               help="Input genome reference in fasta format")
+@click.option("-g",
+              "--genome-file",
+              'genome_file',
+              required=True,
+              type=click.Path(),
+              help="Input genome file of reference")
 @click.option("-o",
-              "--output_file",
+              "--output-file",
+              'output_file',
               #default="samples.tsv",
               help="""path to output file; \ndefault:
                       ./[positive, negative]_samples.tsv""")
@@ -29,7 +38,7 @@ from negative_training_sampler.negative_training_balancer import balance_trainin
 @click.option("--memory",
               default="2GB",
               help="""amount of memory per core (e.g. 2 cores * 2GB = 4GB)\ndefault: 2GB""")
-def cli(label_file, genome_file, output_file, verbose, cores, memory):
+def cli(label_file, reference_file, genome_file, output_file, verbose, cores, memory):
     # pylint: disable=no-value-for-parameter
     '''
     A simple script that takes a tsv file with positive and negative labels
@@ -38,6 +47,7 @@ def cli(label_file, genome_file, output_file, verbose, cores, memory):
     '''
 
     balance_trainingdata(label_file=label_file,
+                         reference_file=reference_file,
                          genome_file=genome_file,
                          output_file=output_file,
                          cores=cores,