+++ negative_training_sampler/io.py
@@ -1,5 +1,7 @@
 #!/usr/bin/env python
 
+import csv
+
 """i/o functions"""
 
 def save_to_file(sample_df, output_file):
@@ -12,3 +14,21 @@ def save_to_file(sample_df, output_file):
 
     with open(output_file, "w") as file:
         sample_df.to_csv(file, sep="\t", index=False, header=False)
+
+
+def load_contigs(genome_file):
+    """loads a genome file of the reference.
+
+    Arguments:
+        genome_file {str} -- [path to the genome file]
+
+    Returns:
+        [list] -- list containing contigs of the genome file
+    """
+
+    contigs = set()
+    with open(genome_file) as file:
+        genome_file_reader = csv.reader(file, delimiter='\t')
+        for contig in genome_file_reader:
+            contigs.add(contig[0])
+    return list(contigs)