+++ negative_training_sampler/negative_training_balancer.py
@@ -8,18 +8,12 @@ from negative_training_sampler.filter import get_negative
 from negative_training_sampler.filter import get_positive
 from negative_training_sampler.filter import clean_sample
 from negative_training_sampler.io import save_to_file
+from negative_training_sampler.io import load_contigs
 from negative_training_sampler.utils import combine_samples
 
-CHROMS = ["chr1", "chr2", "chr3",
-          "chr4", "chr5", "chr6",
-          "chr7", "chr8", "chr9",
-          "chr10", "chr11", "chr12",
-          "chr13", "chr14", "chr15",
-          "chr16", "chr17", "chr18",
-          "chr19", "chr20", "chr21",
-          "chr22", "chrX", "chrY"]
 
 def balance_trainingdata(label_file,
+                         reference_file,
                          genome_file,
                          output_file,
                          verbose,
@@ -32,7 +26,8 @@ def balance_trainingdata(label_file,
     Arguments:
         label_file {str}        -- [Path to a tab separated file containing genomic regions
                                     labeled as positive(1) or negative(0)]
-        genome_file {str}       -- [Path to a refrence genome in FASTA format]
+        reference_file {str}    -- [Path to a reference genome in FASTA format]
+        genome_file {str}       -- [Path to the genome file of the reference]
         output_file {str}       -- [Name of the output file. File will be in .tsv format.]
         cores {int}             -- [Number of cores, default is 1. ]
         memory_per_core {str}   -- [Amount of memory per core.
@@ -51,7 +46,7 @@ def balance_trainingdata(label_file,
     if verbose:
         print("---------------------\ncalculating GC content...\n---------------------")
 
-    cl_gc = get_gc(label_file, genome_file)
+    cl_gc = get_gc(label_file, reference_file)
 
     if verbose:
         print("---------------------\nextracting positive samples...\n---------------------")
@@ -67,10 +62,15 @@ def balance_trainingdata(label_file,
                        ).compute()
 
     if verbose:
+        print("---------------------\nloading contigs...\n---------------------")
+
+    contigs = load_contigs(genome_file)
+
+    if verbose:
         print("---------------------\ncleaning samples\n---------------------")
 
-    positive_sample_cleaned = clean_sample(positive_sample, CHROMS)
-    negative_sample_cleaned = clean_sample(negative_sample, CHROMS)
+    positive_sample_cleaned = clean_sample(positive_sample, contigs)
+    negative_sample_cleaned = clean_sample(negative_sample, contigs)
 
     if verbose:
         print("---------------------\nsaving results\n---------------------")