+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## development
 
+* rename genome_file to reference-file
+* use a genome file of the fasta for the contigs by `--genome-file` (see #7) to avoid hard-coded contigs. Now this sampler should work with every species.
 
 
 ## v0.1.0